Refactor response type imports and update enum values in the database for consistency
