+++ prisma/schema.prisma
@@ -45,9 +45,10 @@ model Account {
 }
 
 model Memory {
-  id        String   @id @default(cuid())
-  userId    String
-  content   String
+  id      String @id @default(cuid())
+  userId  String
+  content String
+
   createdAt DateTime @default(now())
   updatedAt DateTime @updatedAt
 
@@ -55,18 +56,19 @@ model Memory {
 }
 
 enum OwnerType {
-  user
-  model
+  USER
+  MODEL
 }
 
 model History {
-  id        String    @id @default(cuid())
-  userId    String
-  content   String
-  owner     OwnerType @default(user)
-  time      DateTime
-  createdAt DateTime  @default(now())
-  updatedAt DateTime  @updatedAt
+  id      String    @id @default(cuid())
+  userId  String
+  content String
+  owner   OwnerType @default(USER)
+  time    DateTime
+
+  createdAt DateTime @default(now())
+  updatedAt DateTime @updatedAt
 
   user User @relation(fields: [userId], references: [id], onDelete: Cascade)
 }
