+++ src/types/response.type.ts
